+++ src/main/java/dev/nullpanic/tamagotchitelegrambot/service/PetConditionService.java
@@ -1,0 +1,8 @@
+package dev.nullpanic.tamagotchitelegrambot.service;
+
+public interface PetConditionService {
+
+    void decreasePetsStatus();
+
+    void sendLowHungrinessPetNotifications();
+}